Renamed to logger
