+++ dbtwiz/admin/cleanup.py
@@ -3,7 +3,7 @@ from dbtwiz.dbt.manifest import Manifest
 from dbtwiz.dbt.target import Target
 from dbtwiz.gcp.auth import ensure_auth
 from dbtwiz.gcp.bigquery import BigQueryClient
-from dbtwiz.helpers.log_types import error, info
+from dbtwiz.helpers.logger import error, info
 from dbtwiz.ui.interact import multiselect_from_list
