+++ dbtwiz/model/from_sql.py
@@ -2,7 +2,7 @@ import re
 from typing import Dict, List, Tuple
 
 from dbtwiz.dbt.manifest import Manifest
-from dbtwiz.helpers.log_types import fatal, info, warn
+from dbtwiz.helpers.logger import fatal, info, warn
 
 
 def build_reference_lookup(manifest: dict) -> Dict[str, Tuple[str, str]]:
